Started with checkifinview script, and made camerafogsettings.
checkifinview checks if the object that that script is on is in the
camera view, we will use this so that we know which interactive objects
can be activated (in your view).

camerafogsettings allows us the change if fog is enabled/ or change the
fog density. this way we can give one camera a better sight than the
other.